requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
aiohttp>=3.8.0